requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-migrate==4.0.7
Flask-login==0.6.3
Flask-wtf==1.2.2
requests
python-dotenv==1.0.1
appbuilder
pdfminer.six==20240706
websocket-client==1.8.0
wtforms
werkzeug
azure-cognitiveservices-speech
azure-storage-blob
